pyproject.toml
[project]
name = "foia-bias-analysis"
version = "0.1.0"
description = "Automated pipeline for measuring partisan bias in FOIA disclosures"
authors = ["Your Team <team@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2",
    "numpy>=1.26",
    "requests>=2.31",
    "pdfplumber>=0.10",
    "pytesseract>=0.3",
    "Pillow>=10.0",
    "beautifulsoup4>=4.12",
    "spacy>=3.7",
    "statsmodels>=0.14",
    "scikit-learn>=1.4",
    "openai>=1.14",
    "python-muckrock>=0.5.0",
    "typer>=0.9",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "pyarrow>=15.0"
]

[project.optional-dependencies]
ocr = ["pytesseract", "Pillow"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
